requirements.txt
alembic==1.0.6
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
Click==7.0
cryptography==2.4.2
cycler==0.10.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-WTF==0.14.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
Mako==1.0.7
MarkupSafe==1.1.0
matplotlib==3.0.2
numpy==1.15.4
pandas==0.23.4
psycopg2==2.7.6.1
pycparser==2.19
pyparsing==2.3.0
PySocks==1.6.8
python-dateutil==2.7.5
python-editor==1.0.3
python-telegram-bot==11.1.0
pytz==2018.9
scipy==1.2.0
seaborn==0.9.0
six==1.12.0
SQLAlchemy==1.2.15
SQLAlchemy-Utils==0.33.9
Werkzeug==0.14.1
WTForms==2.2.1
